add tests for hamburger menu behaviour
