+++ src/lib/navbar/test/BasicNavbar.svelte
@@ -4,11 +4,16 @@
 	import NavItem from '../NavItem.svelte';
 </script>
 
-<NavBar>
-	<a href="https://peopleplus.co.uk" slot="start"> PeoplePlus </a>
-	<NavBarNav slot="end">
-		<NavItem href="/">Home</NavItem>
-		<NavItem href="/about">About</NavItem>
-		<NavItem href="/contact">Contact</NavItem>
-	</NavBarNav>
-</NavBar>
+<div data-testid="navbar">
+	<NavBar>
+		<a data-testid="start" slot="start" href="https://peopleplus.co.uk"> PeoplePlus </a>
+		<input data-testid="middle" slot="middle" type="search" placeholder="Search" />
+		<div data-testid="end" slot="end">
+			<NavBarNav>
+				<NavItem href="/">Home</NavItem>
+				<NavItem href="/about">About</NavItem>
+				<NavItem href="/contact">Contact</NavItem>
+			</NavBarNav>
+		</div>
+	</NavBar>
+</div>
