+++ src/routes/(docs)/Accordion/ExampleAccordion.svelte
