+++ src/routes/(docs)/Carousel/ExampleCarousel.svelte
