+++ src/routes/(docs)/CookieBanner/+page.svelte
