+++ src/routes/(docs)/DarkMode/ExampleDarkMode.svelte
