+++ src/routes/(docs)/ExamplePage.svelte
