+++ src/routes/(docs)/Footer/+page.svelte
