+++ src/routes/(docs)/Jumbotron/ExampleJumbotron.svelte
