+++ src/routes/(docs)/NavBar/ExampleNavBar.svelte
