+++ src/routes/(docs)/Stepper/+page.svelte
