+++ src/routes/(docs)/Stepper/ExampleStepper.svelte
