+++ src/routes/(docs)/guidelines/+page.svelte
