+++ src/routes/test/navbar/+page.svelte
@@ -1,0 +1,5 @@
+<script>
+	import BasicNavbar from '$lib/navbar/test/BasicNavbar.svelte';
+</script>
+
+<BasicNavbar />
